% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg-package.R
\name{transpose}
\alias{transpose}
\title{Transpose a matrix}
\usage{
transpose(X)
}
\arguments{
\item{X}{numeric matrix}
}
\value{
numeric matrix
}
\description{
Transpose a matrix
}
\examples{
set.seed(1234)
X <- matrix(rnorm(4), nrow = 2, ncol = 2)
X
transpose(X)
}
